split up flow type defs
