+++ src/internalTypes/founddation/entityState.js
@@ -1,0 +1,8 @@
+// @flow
+export type EntityCollectionT = {
+  [entityId: string]: any,
+}
+
+export type EntitiesState = {
+  [collection: string]: EntityCollectionT,
+}
